feat: upgrade Chart.js to v4
Chart.js v4 support

BREAKING CHANGE: package now is ESM-only, no CommonJS support
