+++ .storybook/manager.js
@@ -1,6 +1,6 @@
 import { addons } from '@storybook/addons'
 
-import { theme } from './theme'
+import { theme } from './theme.js'
 
 addons.setConfig({
   theme,
