+++ sandboxes/bar/package.json
@@ -1,11 +1,12 @@
 {
+  "type": "module",
   "scripts": {
     "start": "vite"
   },
   "dependencies": {
-    "chart.js": "^3.8.0",
+    "chart.js": "^4.0.0",
     "vue": "^3.2.31",
-    "vue-chartjs": "^4.0.0"
+    "vue-chartjs": "^5.0.0"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^3.0.1",
