+++ sandboxes/scatter/src/App.vue
@@ -12,7 +12,7 @@ import {
   Legend
 } from 'chart.js'
 import { Scatter } from 'vue-chartjs'
-import * as chartConfig from './chartConfig'
+import * as chartConfig from './chartConfig.js'
 
 ChartJS.register(LinearScale, PointElement, LineElement, Tooltip, Legend)
