+++ stories/radar.stories.ts
@@ -1,6 +1,6 @@
 import 'chart.js/auto'
-import { Radar } from '../src'
-import * as radarChartConfig from '../sandboxes/radar/src/chartConfig'
+import { Radar } from '../src/index.js'
+import * as radarChartConfig from '../sandboxes/radar/src/chartConfig.js'
 
 export default {
   title: 'RadarChart',
