+++ tsconfig.json
@@ -10,12 +10,13 @@
     "noUnusedParameters": true,
     /* Modules */
     "baseUrl": ".",
-    "module": "esnext",
-    "moduleResolution": "node",
+    "module": "ESNext",
+    "moduleResolution": "Node", // restore to "NodeNext" after Chart.js update
     "resolveJsonModule": true,
     /* Emit */
     "declaration": true,
     "declarationMap": true,
+    "importsNotUsedAsValues": "error",
     "inlineSourceMap": true,
     "outDir": "dist",
     /* Interop Constraints */
@@ -26,7 +27,7 @@
       "dom",
       "esnext"
     ],
-    "target": "esnext",
+    "target": "ESNext",
     /* Completeness */
     "skipLibCheck": true
   },
